rearrange rocksdb backend into its own module
