+++ rocks/tests/test_backend.py
@@ -1,32 +1,35 @@
 import os
+import shutil
+import tempfile
 import unittest
-from unittest import skipIf
 
 from polymr.record import Record
-import polymr_postgres
-
-ENVVAR = "POLYMR_POSTGRES_URL"
-URL = os.environ.get(ENVVAR, False)
-should_skip_test = not bool(URL)
+import polymr.storage
+import polymr_rocksdb
 
 
-class TestPostgresBackend(unittest.TestCase):
+polymr_rocksdb  # pyflakes
+
+
+class TestRocksDBBackend(unittest.TestCase):
+
     def setUp(self):
-        self.db = polymr_postgres.PostgresBackend(URL)
+        self.workdir = tempfile.mkdtemp(suffix="polymrtest")
+        self.db = None
 
     def tearDown(self):
-        self.db.destroy()
-        self.db.close()
+        if os.path.exists(self.workdir):
+            shutil.rmtree(self.workdir)
 
     def _get_db(self, new=False):
         if self.db and not new:
             return self.db
         if self.db:
             self.db.close()
-        self.db = polymr_postgres.PostgresBackend(URL)
+        self.db = polymr.storage.parse_url(
+            "rocksdb://localhost"+self.workdir)
         return self.db
 
-    @skipIf(should_skip_test, ENVVAR+" not defined")
     def test_get_set_freqs(self):
         db = self._get_db()
         x = {b"abc": 3, b"bcd": 2}
@@ -34,7 +37,13 @@ class TestPostgresBackend(unittest.TestCase):
         y = db.get_freqs()
         self.assertEqual(x, dict(y))
 
-    @skipIf(should_skip_test, ENVVAR+" not defined")
+    def test_get_set_rowcount(self):
+        db = self._get_db()
+        x = 222
+        db.save_rowcount(x)
+        y = db.get_rowcount()
+        self.assertEqual(x, y)
+
     def test_get_set_token(self):
         db = self._get_db()
         tok = b"abc"
@@ -49,15 +58,14 @@ class TestPostgresBackend(unittest.TestCase):
         self.assertNotEqual(records_cmpct, rng)
         self.assertEqual(rng, list(range(1,8)))
 
-    @skipIf(should_skip_test, ENVVAR+" not defined")
     def test_get_set_records(self):
         db = self._get_db()
         r1 = Record(["abcde", "foo"], "1", ['dogsays'])
         r2 = Record(["qwert", "bar"], "2", ['barque'])
-        db.save_records(enumerate((r1, r2)))
+        cnt = db.save_records(enumerate((r1, r2)))
+        self.assertEqual(cnt, 2)
         r1_db, r2_db = list(db.get_records([0, 1]))
         self.assertEqual(r1.fields, r1_db.fields)
         self.assertEqual(r1.pk, r1_db.pk)
         self.assertEqual(r2.fields, r2_db.fields)
         self.assertEqual(r2.pk, r2_db.pk)
-
