+++ rocks/tests/test_endtoend.py
@@ -8,6 +8,10 @@ import polymr.index
 import polymr.storage
 import polymr.query
 import polymr.record
+import polymr_rocksdb
+
+
+polymr_rocksdb  # pyflakes
 
 to_index = StringIO("""01001,MA,DONNA,AGAWAM,WUCHERT,PO BOX 329,9799PNOVAY
 01007,MA,BERONE,BELCHERTOWN,BOARDWAY,135 FEDERAL ST,9799JA8CB5
@@ -25,13 +29,12 @@ sample_query = ["01030","MELANI","PICKETT","18 PAUL REVERE DR"]
 sample_pk = "989960D48D"
 
 
-class TestEndToEnd(unittest.TestCase):
+class TestEndToEndWithRocksDB(unittest.TestCase):
     def setUp(self):
         self.workdir = tempfile.mkdtemp(suffix="polymrtest")
         self.db = polymr.storage.parse_url(
-            "leveldb://localhost"+self.workdir)
-        to_index.seek(0)
-
+            "rocksdb://localhost"+self.workdir)
+        
     def tearDown(self):
         self.db.close()
         del self.db
@@ -60,46 +63,5 @@ class TestEndToEnd(unittest.TestCase):
         self.assertEqual(hit['pk'], sample_pk,"searches should survive typos")
         
 
-class TestEndToEndParallel(unittest.TestCase):
-    def setUp(self):
-        self.workdir = tempfile.mkdtemp(suffix="polymrtest")
-        self.url = "leveldb://localhost"+self.workdir
-        to_index.seek(0)
-
-    def tearDown(self):
-        if os.path.exists(self.workdir):
-            shutil.rmtree(self.workdir)
-        to_index.seek(0)
-
-    def test_end_to_end(self):
-        recs = polymr.record.from_csv(
-            to_index,
-            searched_fields_idxs=[0,2,4,5],
-            pk_field_idx=-1,
-            include_data=False
-        )
-        db = polymr.storage.parse_url(self.url)
-        polymr.index.create(recs, 1, 10, db)
-        del db
-        index = polymr.query.ParallelIndex(self.url, 2)
-        hit = index.search(sample_query, limit=1)[0]
-        self.assertEqual(hit['pk'], sample_pk,
-                         ("querying the index with an indexed record should "
-                          "return that same record"))
-        
-        tpyo1 = list(sample_query[0])
-        tpyo1[2], tpyo1[3] = tpyo1[3], tpyo1[2]
-        tpyo1 = "".join(tpyo1)
-        tpyo2 = list(sample_query[0])
-        tpyo2[1], tpyo2[2] = tpyo2[2], tpyo2[1]
-        tpyo2 = "".join(tpyo2)
-        results = index.searchmany([[tpyo1]+sample_query[1:],
-                                    [tpyo2]+sample_query[1:]], limit=1)
-        for result in results:
-            hit = result[0]
-            self.assertEqual(hit['pk'], sample_pk,
-                             "searches should survive typos")
-        
-        
 if __name__ == '__main__':
     unittest.main()
